refactor: ShapeDrawingEngin을 interface에서 abstract로 변경
